backend/requirements.txt
Django==2.1.2
djangorestframework==3.9.0
django-filter==2.0.0
django-rest-knox==3.4.0
pytz==2018.6
